requirements.txt
PyGithub
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
requests
colored
